Deep scan error fixed
