+++ package.json
@@ -53,6 +53,7 @@
 		"react-navigation-stack": "^2.5.1",
 		"react-navigation-tabs": "^2.8.13",
 		"react-redux": "^7.2.0",
+		"realm": "^10.0.0-beta.6",
 		"redux": "^4.0.5",
 		"redux-logger": "^3.0.6",
 		"redux-saga": "^1.1.3",
